+++ examples/snippets/widgets/sevenseg.h
@@ -45,4 +45,4 @@
 #ifndef EVE_OPT_FILL
 #define EVE_OPT_FILL 0x2000ul
 #endif
-void sevensegment(int32_t x, int32_t y, uint16_t size, uint16_t digit, uint32_t fgcolour, uint32_t bgcolour);
+void sevensegment(int16_t x, int16_t y, uint16_t size, uint16_t digit, uint32_t fgcolour, uint32_t bgcolour);
